+++ lib/vfpage/src/types.ts
@@ -1,11 +1,11 @@
-export type CreateLwcInput = {
+export interface CreateLwcInput {
   vfns: string;
   ns: string;
   component: string;
   props: any;
   methods: string[] | null | undefined;
   refs: string[] | null | undefined;
-};
+}
 
 export type MobileMethod = {[K: string]: (...a: any) => Promise<any>};
 
@@ -20,15 +20,15 @@ export type CallbackInput = [responseData: any, error: string | null];
 
 export type CallbackMap = Map<string, (...props: CallbackInput) => void>;
 
-export type PostMessageData = {
+export interface PostMessageData {
   data: any;
   type: RequestType;
   id: string;
-};
+}
 
 export type LwcMobile = HTMLElement & {mobileMethods: MobileMethod | null};
 
-export type SetChildrenOptions = {
+export interface SetChildrenOptions {
   element: string;
   props: any;
-};
+}