+++ react-native-demo/config.js
@@ -1,0 +1,19 @@
+const config = {
+  callbackUrl: 'vloc://vloc/success',
+  clientId: '3MVG9CEn_O3jvv0x_GqVPvcLpo7KO.USvG5g9_oPpSd7ZQLY_p.aa.5hioTv9rPljrTYmryejDVHjNIKdGrAv',
+  clientSecret: 'C538CF52EDF142F10A435A17A63677BC9E9F62F7A28C5AA32DFFE1D20F389D7D',
+  authUrl: 'https://login.salesforce.com/services/oauth2/authorize',
+  apiVersion: 'v45.0',
+  responseType: 'token'
+};
+
+//smtp org
+// const config = {
+//   callbackUrl: 'vloc://vloc/success',
+//   clientId: '3MVG9Iu66FKeHhINtY8g_aW0NhtqfnlxvKJ2gsMPZlROwze3OJUoBl8BnRS6UnqpqCtdYmFmBshCjSCjqxSa2',
+//   authUrl: 'https://login.stmpa.stm.salesforce.com/services/oauth2/authorize',
+//   apiVersion: 'v45.0',
+//   responseType: 'token'
+// };
+
+export default config;